Adds simple example for checking httpstatus with DIO
As requested by the community I'm adding a simple example of making a
call with DIO and getting the HTTP Response code to validate if we have
to upgrade or not.

DIO has a particular way for handling exceptions [1] that needs to by
try catch to check status codes that are considered errors.

Closes #18

[1]: https://pub.dev/packages/dio#handling-errors