Improves ability to translate/customise messages
As the community needs more flexibility for customising the language
this package provides by default.

New fields are not added to cover the hard coded messages which will
also allow for translations if the user decides to do that too. It could
have a better way to delegate the language but it's ok to have it as
parameters for now.